# docs/swagger.yaml
openapi: 3.0.3
info:
  title: TeamSupportPro API
  description: REST API for TeamSupportPro — a team-based IT support ticket management system.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: team_support_session
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ActivityEntry:
      type: object
      properties:
        id: { type: string }
        ticketId: { type: string }
        actor: { type: string }
        message: { type: string }
        at: { type: string, format: date-time }
    Ticket:
      type: object
      properties:
        id: { type: string, example: "TKT-10001" }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [Open, In Progress, Pending, Resolved, Closed] }
        priority: { type: string, enum: [Low, Medium, High, Critical] }
        teamId: { type: string }
        categoryId: { type: string }
        assignedToId: { type: string, nullable: true }
        requestorName: { type: string }
        requestorEmail: { type: string }
        location: { type: string }
        dueLabel: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        activity:
          type: array
          items: { $ref: '#/components/schemas/ActivityEntry' }
    TicketAttachment:
      type: object
      properties:
        id: { type: string }
        ticketId: { type: string }
        fileName: { type: string }
        contentType: { type: string }
        fileSizeBytes: { type: integer }
        uploadedByUserId: { type: string }
        uploadedByName: { type: string }
        uploadedAt: { type: string, format: date-time }
...
security:
  - cookieAuth: []
  - apiKeyHeader: []
  - bearerAuth: []
paths:
...
  /api/tickets:
    get:
      summary: List tickets for the authenticated user's team
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items: { $ref: '#/components/schemas/Ticket' }
        "401":
          description: Unauthenticated
    post:
      summary: Create a new ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, priority, teamId, categoryId, requestorName, requestorEmail]
              properties:
                id: { type: string }
                title: { type: string }
                description: { type: string }
                priority: { type: string, enum: [Low, Medium, High, Critical] }
                teamId: { type: string }
                categoryId: { type: string }
                assignedToId: { type: string, nullable: true }
                requestorName: { type: string }
                requestorEmail: { type: string }
                location: { type: string }
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket: { $ref: '#/components/schemas/Ticket' }
        "400":
          description: Invalid input
        "401":
          description: Unauthenticated
        "403":
          description: Cross-team ticket creation forbidden
  /api/tickets/{ticketId}:
    get:
      summary: Get a ticket by ID
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket: { $ref: '#/components/schemas/Ticket' }
        "404":
          description: Ticket not found
    patch:
      summary: Update a ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                status: { type: string, enum: [Open, In Progress, Pending, Resolved, Closed] }
                priority: { type: string, enum: [Low, Medium, High, Critical] }
                categoryId: { type: string }
                assignedToId: { type: string, nullable: true }
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket: { $ref: '#/components/schemas/Ticket' }
        "400":
          description: Invalid input
        "403":
          description: Cross-team update forbidden
    delete:
      summary: Delete a ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Ticket deleted
        "403":
          description: Cross-team delete forbidden
        "404":
          description: Ticket not found
  /api/tickets/{ticketId}/comments:
    post:
      summary: Add a comment to a ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment: { $ref: '#/components/schemas/ActivityEntry' }
        "400":
          description: Invalid payload
        "403":
          description: Cross-team comment forbidden
  /api/tickets/{ticketId}/attachments:
    get:
      summary: List attachments for a ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of attachments
          content:
            application/json:
              schema:
                type: object
                properties:
                  attachments:
                    type: array
                    items: { $ref: '#/components/schemas/TicketAttachment' }
    post:
      summary: Upload an attachment
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Attachment uploaded
        "400":
          description: Missing file or invalid
  /api/tickets/{ticketId}/attachments/{attachmentId}:
    get:
      summary: Download an attachment
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
        - name: attachmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Attachment not found
    delete:
      summary: Delete an attachment (soft delete)
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
        - name: attachmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Attachment deleted
        "404":
          description: Attachment not found
...
  /api/public/tickets:
    post:
      summary: Create an anonymous/public ticket
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId, categoryId, title, description, requestorName, requestorEmail]
              properties:
                teamId: { type: string }
                categoryId: { type: string }
                title: { type: string }
                description: { type: string }
                requestorName: { type: string }
                requestorEmail: { type: string }
                location: { type: string }
      responses:
        "201":
          description: Public ticket created
        "400":
          description: Invalid input
...